change trigger_count to unsigned long; change prints to say 'instructions'

+++ work/modules/landslide/landslide.h
@@ -21,8 +21,8 @@ struct ls_state {
 	/* log_object_t must be the first thing in the device struct */
 	log_object_t log;
 
-	int trigger_count; /* in this branch of the tree */
-	int absolute_trigger_count; /* in the whole execution */
+	unsigned long trigger_count; /* in this branch of the tree */
+	unsigned long absolute_trigger_count; /* in the whole execution */
 
 	/* Pointers to relevant objects. Currently only supports one CPU. */
 	conf_object_t *cpu0;
